enforce all-named arguments on arrows2d
